--- a/packages/frontend/package.json
+++ b/packages/frontend/package.json
@@ -10,6 +10,7 @@
     "@guardian/pasteup": "1.0.0-alpha.1",
     "aws-sdk": "^2.340.0",
     "compose-function": "^3.0.3",
+    "compression": "^1.7.3",
     "curlyquotes": "^1.3.2",
     "dateformat": "^3.0.3",
     "diskusage": "^0.2.5",
@@ -32,6 +33,7 @@
   "devDependencies": {
     "@types/amphtml-validator": "^1.0.0",
     "@types/compose-function": "^0.0.30",
+    "@types/compression": "^0.0.36",
     "@types/dateformat": "^1.0.1",
     "@types/dompurify": "^0.0.31",
     "@types/express": "^4.16.0",