This is a basic optimisation. I'm still establishing the exact performance increase.

## Without Gzip (CODE, 1 * t3.medium)

```
$ vegeta attack -targets=targets.txt -duration=300s -rate=50 -header="Content-Type: application/json" | tee results.bin | vegeta report
Requests      [total, rate]            15000, 50.00
Duration      [total, attack, wait]    5m0.49418077s, 4m59.990233s, 503.94777ms
Latencies     [mean, 50, 95, 99, max]  1.581154392s, 702.112041ms, 5.956441276s, 10.247213956s, 42.455166734s
Bytes In      [total, mean]            934751070, 62316.74
Bytes Out     [total, mean]            2025095130, 135006.34
Success       [ratio]                  98.20%
Status Codes  [code:count]             0:270  200:14730  
```

### With GZIP

```
$ vegeta attack -targets=targets-code.txt -duration=300s -rate=50 -header="Content-Type: application/json" | tee results.bin | vegeta report
Requests      [total, rate]            15000, 50.00
Duration      [total, attack, wait]    5m0.036254486s, 4m59.989438s, 46.816486ms
Latencies     [mean, 50, 95, 99, max]  437.999337ms, 57.07538ms, 3.864362122s, 6.201623375s, 9.105618054s
Bytes In      [total, mean]            943989288, 62932.62
Bytes Out     [total, mean]            2042417736, 136161.18
Success       [ratio]                  98.91%
Status Codes  [code:count]             0:144  200:14837  504:19  
```

This is an improvement, particularly in latency - the mean is ⅓ of before (1.5s - .5s), still too high but much better. Errors have fallen from 270 to 163 (40% fall).